requirements.txt
flask
flask_login
flask_migrate
flask_admin
flask_wtf
flask_sqlalchemy==2.4.1
sqlalchemy==1.3.17
email_validator
python-decouple
gunicorn
quandl
pandas
plotly
xlrd
openpyxl
numpy
html5lib
yahoofinancials
yfinance
psycopg2
pyEX
currency_symbols
git+git://github.com/GeneralMills/pytrends.git